chore: folders refactoring
