+++ src/components/Month.tsx
@@ -1,6 +1,6 @@
 import { QRL, component$ } from '@builder.io/qwik';
 import { t } from '../locale/labels';
-import { Month as TMonth } from '../utils/types';
+import { Month as TMonth } from '../models/types';
 
 export const Month = component$<{
 	month: TMonth;
