+++ src/pages/Skills.tsx
@@ -4,8 +4,9 @@ import { Filters } from '../components/Filters';
 import { SkillCard } from '../components/SkillCard';
 import { SkillLegend } from '../components/SkillLegend';
 import { tt } from '../locale/labels';
-import { getConfiguration, getSkills } from '../utils/api';
-import { SkillMatrix } from '../utils/types';
+import { SkillMatrix } from '../models/types';
+import { getConfiguration } from '../services/configuration';
+import { getSkills } from '../services/skillMatrix';
 
 export const Skills = component$(() => {
 	const appStore = useContext(AppContext);
