+++ src/router.tsx
@@ -1,10 +1,10 @@
 import { Signal, useSignal, useTask$ } from '@builder.io/qwik';
 import { Auth } from './components/Auth';
-import { Effort } from './page/Effort';
-import { Profile } from './page/Profile';
-import { Search } from './page/Search';
-import { Skills } from './page/Skills';
-import { Timesheet } from './page/Timesheet';
+import { Effort } from './pages/Effort';
+import { Profile } from './pages/Profile';
+import { Search } from './pages/Search';
+import { Skills } from './pages/Skills';
+import { Timesheet } from './pages/Timesheet';
 
 export type Route = keyof typeof routes;
