+++ src/services/projects.ts
@@ -1,0 +1,5 @@
+import { Customer, Project } from '../models/types';
+import { getHttpResponse } from '../network/httpRequest';
+
+export const getProjects = async (company: string = 'it', customer: Customer): Promise<Project[]> =>
+	getHttpResponse<Project[]>(`task/project?company=${company}&customer=${customer}`);
