+++ src/services/tasks.ts
@@ -1,0 +1,9 @@
+import { getHttpResponse } from '../network/httpRequest';
+import { Customer, Project, Task } from '../models/types';
+
+export const getTasks = async (
+	company: string = 'it',
+	customer: Customer,
+	project: Project
+): Promise<Task[]> =>
+	getHttpResponse<Task[]>(`task/task?company=${company}&customer=${customer}&project=${project}`);
